Let the diagonalizers support different floating point precisions

+++ src/cpp/src/basis/BasisAtomCreator.cpp
@@ -21,7 +21,7 @@ BasisAtomCreator<Scalar> &BasisAtomCreator<Scalar>::restrict_energy(real_t min, 
 template <typename Scalar>
 BasisAtomCreator<Scalar> &BasisAtomCreator<Scalar>::restrict_quantum_number_f(real_t min,
                                                                               real_t max) {
-    if (2 * min != std::rintf(2 * min) || 2 * max != std::rintf(2 * max)) {
+    if (2 * min != std::rint(2 * min) || 2 * max != std::rint(2 * max)) {
         throw std::invalid_argument("Quantum number f must be an integer or half-integer.");
     }
     range_quantum_number_f = {min, max};
@@ -31,7 +31,7 @@ BasisAtomCreator<Scalar> &BasisAtomCreator<Scalar>::restrict_quantum_number_f(re
 template <typename Scalar>
 BasisAtomCreator<Scalar> &BasisAtomCreator<Scalar>::restrict_quantum_number_m(real_t min,
                                                                               real_t max) {
-    if (2 * min != std::rintf(2 * min) || 2 * max != std::rintf(2 * max)) {
+    if (2 * min != std::rint(2 * min) || 2 * max != std::rint(2 * max)) {
         throw std::invalid_argument("Quantum number m must be an integer or half-integer.");
     }
     range_quantum_number_m = {min, max};
